Separate modal into own file and refactor handle function
