+++ frontend/beCompliant/src/components/table/Comment.tsx
@@ -1,14 +1,7 @@
 import {
-  Button,
   Center,
   HStack,
   IconButton,
-  Modal,
-  ModalBody,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
   Spacer,
   Stack,
   Text,
@@ -16,71 +9,8 @@ import {
   useDisclosure,
 } from '@kvib/react';
 import { useState } from 'react';
-import { useDeleteComment } from '../../hooks/useDeleteComment';
 import { useSubmitComment } from '../../hooks/useSubmitComment';
-
-type DeleteCommentModalProps = {
-  onOpen: () => void;
-  onClose: () => void;
-  isOpen: boolean;
-  comment: string;
-  questionId: string;
-  team: string | undefined;
-  setEditMode: (value: boolean) => void;
-};
-function DeleteCommentModal({
-  onOpen,
-  onClose,
-  isOpen,
-  comment,
-  questionId,
-  team,
-  setEditMode,
-}: DeleteCommentModalProps) {
-  const { mutate: deleteComment, isPending: isLoading } =
-    useDeleteComment(setEditMode);
-  const handleCommentDelete = () => {
-    deleteComment({
-      actor: 'Unknown',
-      questionId: questionId,
-      team: team,
-      comment: comment,
-      updated: new Date(),
-    });
-    setEditMode(false);
-  };
-
-  return (
-    <Modal onClose={onClose} isOpen={isOpen} isCentered>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>Slett kommentar</ModalHeader>
-        <ModalBody>
-          <Stack>
-            <Text size="sm">Er du sikker på at du vil slette kommentaren?</Text>
-          </Stack>
-        </ModalBody>
-        <ModalFooter>
-          <HStack justifyContent="end">
-            <Button variant="tertiary" colorScheme="blue" onClick={onClose}>
-              Avbryt
-            </Button>
-            <Button
-              aria-label="Slett kommentar"
-              variant="primary"
-              colorScheme="red"
-              leftIcon="delete"
-              onClick={handleCommentDelete}
-              isLoading={isLoading}
-            >
-              Slett kommentar
-            </Button>
-          </HStack>
-        </ModalFooter>
-      </ModalContent>
-    </Modal>
-  );
-}
+import { DeleteCommentModal } from './DeleteCommentModal';
 
 // Replace with type from api when the internal data model is implemented
 type CommentProps = {
@@ -114,12 +44,6 @@ export function Comment({ comment, questionId, team }: CommentProps) {
     }
   };
 
-  const handleEditedCommentState = (
-    e: React.ChangeEvent<HTMLTextAreaElement>
-  ) => {
-    setEditedComment(e.target.value);
-  };
-
   const handleDiscardChanges = () => {
     setEditedComment(comment);
     setEditMode(false);
@@ -132,7 +56,7 @@ export function Comment({ comment, questionId, team }: CommentProps) {
           marginBottom={2}
           marginTop={2}
           defaultValue={editedComment}
-          onChange={handleEditedCommentState}
+          onChange={(e) => setEditedComment(e.target.value)}
           size="sm"
         />
         <Spacer />
